moving switch to layout + styles refactors
