+++ apps/google-docs/src/locations/Page/components/overview/OverviewSection.tsx
@@ -1,25 +1,16 @@
 import { useMemo } from 'react';
 import { Box, Button, Flex, Note, Paragraph, Text } from '@contentful/f36-components';
-import { EyeIcon, LightbulbIcon, PencilSimpleIcon } from '@contentful/f36-icons';
-import { cx } from '@emotion/css';
+import { LightbulbIcon } from '@contentful/f36-icons';
 import type { MappingReviewSuspendPayload } from '@types';
 import { buildEntryListFromEntryBlockGraph } from '../../../../utils/overviewEntryList';
 import { OverviewEntryList } from './OverviewEntryList';
-import {
-  modeToggleButton,
-  modeToggleButtonActive,
-  modeToggleWrapper,
-  overviewSectionBox,
-  overviewSectionBoxScrollable,
-} from './OverviewSection.styles';
+import { overviewSectionBox, overviewSectionBoxScrollable } from './OverviewSection.styles';
 import Splitter from '../mainpage/Splitter';
 
 interface OverviewProps {
   payload: MappingReviewSuspendPayload;
   selectedEntryIndex: number | null;
   onSelectEntryIndex: (index: number) => void;
-  reviewMode: 'view' | 'edit';
-  onReviewModeChange: (mode: 'view' | 'edit') => void;
   ctaLabel: string;
   onCtaClick: () => void;
   isCtaLoading?: boolean;
@@ -29,8 +20,6 @@ const OverviewSection = ({
   payload,
   selectedEntryIndex,
   onSelectEntryIndex,
-  reviewMode,
-  onReviewModeChange,
   ctaLabel,
   onCtaClick,
   isCtaLoading = false,
@@ -71,24 +60,6 @@ const OverviewSection = ({
             </Flex>
 
             <Flex alignItems="center" gap="spacingS">
-              <div className={modeToggleWrapper} role="group" aria-label="Review mode">
-                <button
-                  type="button"
-                  className={cx(modeToggleButton, reviewMode === 'view' && modeToggleButtonActive)}
-                  onClick={() => onReviewModeChange('view')}
-                  aria-pressed={reviewMode === 'view'}>
-                  <EyeIcon size="small" />
-                  View only
-                </button>
-                <button
-                  type="button"
-                  className={cx(modeToggleButton, reviewMode === 'edit' && modeToggleButtonActive)}
-                  onClick={() => onReviewModeChange('edit')}
-                  aria-pressed={reviewMode === 'edit'}>
-                  <PencilSimpleIcon size="small" />
-                  Edit mode
-                </button>
-              </div>
               <Button
                 variant="primary"
                 onClick={onCtaClick}
